(PC-35244)[API] refactor: split offers and categories test files
